back/requirements.txt
Flask
psycopg2-binary
python-dotenv
Flask-CORS